Don't redisplay vendor logo with 4.19+ kernels
Hans de Goede explains:

"
Some hardware does not show the vendor logo unless we put it
there ourselves, so the EFIFB driver always "restores" the logo
in case it is not there, this avoids the user just staring at
a black screen all the way into gdm.

This also helps the user not looking at an empty screen for
circa 6 seconds after GRUB when not using LUKS.

Starting with kernel 4.19 you can add: "video=efifb:nobgrt" to
the kernel commandline to explicitly disable this.
"